Move dependency graph functions into dependency_graph.py
