+++ flowmachine/flowmachine/utils.py
@@ -8,21 +8,14 @@ Various simple utilities.
 """
 
 import datetime
-import logging
 import re
 from contextlib import contextmanager
-import networkx as nx
 import structlog
-import sys
-from io import BytesIO
 from pathlib import Path
 from pglast import prettify
 from psycopg2._psycopg import adapt
 from time import sleep
-from typing import List, Union, Tuple, Dict, Sequence, Callable, Any
-from flowmachine.core.errors import UnstorableQueryError
-
-logger = structlog.get_logger("flowmachine.debug", submodule=__name__)
+from typing import Union, Tuple
 
 
 def parse_datestring(
@@ -323,260 +316,3 @@ def sort_recursively(d):
         return sorted(d, key=to_nested_list)
     else:
         return d
-
-
-def print_dependency_tree(query_obj, show_stored=False, stream=None, indent_level=0):
-    """
-    Print the dependencies of a flowmachine query in a tree-like structure.
-
-    Parameters
-    ----------
-    query_obj : Query
-        An instance of a query object.
-    show_stored : bool, optional
-        If True, show for each query whether it is stored or not. Default: False.
-    stream : io.IOBase, optional
-        The stream to which the output should be written (default: stdout).
-    indent_level : int
-        The current level of indentation.
-    """
-    stream = stream or sys.stdout
-    indent_level = indent_level or 0
-
-    indent_per_level = 3
-    indent = " " * (indent_per_level * indent_level - 1)
-    prefix = "" if indent_level == 0 else "- "
-    fmt = "query_id" if not show_stored else "query_id,is_stored"
-    stream.write(f"{indent}{prefix}{query_obj:{fmt}}\n")
-    deps_sorted_by_query_id = sorted(query_obj.dependencies, key=lambda q: q.md5)
-    for dep in deps_sorted_by_query_id:
-        print_dependency_tree(
-            dep, indent_level=indent_level + 1, stream=stream, show_stored=show_stored
-        )
-
-
-def _get_query_attrs_for_dependency_graph(query_obj, analyse=False):
-    """
-    Helper method which returns information about this query for use in a dependency graph.
-
-    Parameters
-    ----------
-    query_obj : Query
-        Query object to return information for.
-    analyse : bool
-        Set to True to get actual runtimes for queries. Note that this will actually run the query!
-
-    Returns
-    -------
-    dict
-        Dictionary containing the keys "name", "stored", "cost" and "runtime" (the latter is only
-        present if `analyse=True`.
-        Example return value: `{"name": "DailyLocation", "stored": False, "cost": 334.53, "runtime": 161.6}`
-    """
-    expl = query_obj.explain(format="json", analyse=analyse)[0]
-    attrs = {}
-    attrs["name"] = query_obj.__class__.__name__
-    attrs["stored"] = query_obj.is_stored
-    attrs["cost"] = expl["Plan"]["Total Cost"]
-    if analyse:
-        attrs["runtime"] = expl["Execution Time"]
-    return attrs
-
-
-def assemble_dependency_graph(
-    dependencies: Sequence[Tuple[Union["Query", None], "Query"]],
-    attrs_func: Callable[["Query"], Dict[str, Any]],
-) -> nx.DiGraph:
-    """
-    Helper function to assemble a dependency graph from a list of dependencies.
-
-    Parameters
-    ----------
-    dependencies : list of tuples (query1, query2) or (None, query2)
-        Dependencies between query objects.
-    attrs_func : function
-        Function that returns a dict of node attributes for a query object.
-    
-    Returns
-    -------
-    networkx.DiGraph
-    """
-    g = nx.DiGraph()
-
-    if dependencies:
-        _, y = zip(*dependencies)
-        for n in set(y):
-            attrs = attrs_func(n)
-            g.add_node(f"x{n.md5}", **attrs)
-
-    for x, y in dependencies:
-        if x is not None:
-            g.add_edge(*[f"x{z.md5}" for z in (x, y)])
-
-    return g
-
-
-def calculate_dependency_graph(query_obj: "Query", analyse: bool = False) -> nx.DiGraph:
-    """
-    Produce a graph of all the queries that go into producing this one, with their estimated
-    run costs, and whether they are stored as node attributes.
-
-    The resulting networkx object can then be visualised, or analysed. When visualised,
-    nodes corresponding to stored queries will be rendered green. See the function
-    `plot_dependency_graph()` for a convenient way of plotting a dependency graph directly
-    for visualisation in a Jupyter notebook.
-
-    The dependency graph includes the estimated cost of the query in the 'cost' attribute,
-    the query object the node represents in the 'query_object' attribute, and with the analyse
-    parameter set to true, the actual running time of the query in the `runtime` attribute.
-
-    Parameters
-    ----------
-    query_obj : Query
-        Query object to produce a dependency graph for.
-    analyse : bool
-        Set to True to get actual runtimes for queries. Note that this will actually run the query!
-
-    Returns
-    -------
-    networkx.DiGraph
-
-    Examples
-    --------
-
-    If you don't want to visualise the dependency graph directly (for example
-    using `plot_dependency_graph()`, you can export it to a .dot file as follows:
-
-    >>> import flowmachine
-    >>> from flowmachine.features import daily_location
-    >>> from networkx.drawing.nx_agraph import write_dot
-    >>> flowmachine.connect()
-    >>> G = daily_location("2016-01-01").dependency_graph()
-    >>> write_dot(G, "daily_location_dependencies.dot")
-    >>> G = daily_location("2016-01-01").dependency_graph(True)
-    >>> write_dot(G, "daily_location_dependencies_runtimes.dot")
-
-    The resulting .dot file then be converted to a .pdf file using the external
-    tool `dot` which comes as part of the [GraphViz](https://www.graphviz.org/) package:<br />
-
-    ```
-    $ dot -Tpdf daily_location_dependencies.dot -o daily_location_dependencies.pdf
-    ```
-
-    [Graphviz]: https://www.graphviz.org/
-
-    Notes
-    -----
-    The queries listed as dependencies are not _guaranteed_ to be
-    used in the actual running of a query, only to be referenced by it.
-    """
-    openlist = [(None, query_obj)]
-    deps = []
-
-    while openlist:
-        y, x = openlist.pop()
-        deps.append((y, x))
-
-        openlist += list(zip([x] * len(x.dependencies), x.dependencies))
-
-    def attrs_func(q):
-        attrs = _get_query_attrs_for_dependency_graph(q, analyse=analyse)
-        attrs["shape"] = "rect"
-        attrs["query_object"] = q
-        attrs["label"] = f"{attrs['name']}. Cost: {attrs['cost']}"
-        if analyse:
-            attrs["label"] += " Actual runtime: {}.".format(attrs["runtime"])
-        if attrs["stored"]:
-            attrs["fillcolor"] = "#b3de69"  # light green
-            attrs["style"] = "filled"
-        return attrs
-
-    return assemble_dependency_graph(dependencies=deps, attrs_func=attrs_func)
-
-
-def plot_dependency_graph(
-    query_obj, analyse=False, format="png", width=None, height=None
-):
-    """
-    Plot a graph of all the queries that go into producing this one (see `calculate_dependency_graph`
-    for more details). This returns an IPython.display object which can be directly displayed in
-    Jupyter notebooks.
-
-    Note that this requires the IPython and pygraphviz packages to be installed.
-
-    Parameters
-    ----------
-    query_obj : Query
-        Query object to plot a dependency graph for.
-    analyse : bool
-        Set to True to get actual runtimes for queries. Note that this will actually run the query!
-    format : {"png", "svg"}
-        Output format of the resulting
-    width : int
-        Width in pixels to which to constrain the image. Note this is only supported for format="png".
-    height : int
-        Height in pixels to which to constrain the image. Note this is only supported for format="png".
-
-    Returns
-    -------
-    IPython.display.Image or IPython.display.SVG
-    """
-    try:  # pragma: no cover
-        from IPython.display import Image, SVG
-    except ImportError:
-        raise ImportError("requires IPython ", "https://ipython.org/")
-
-    try:  # pragma: no cover
-        import pygraphviz
-    except ImportError:
-        raise ImportError(
-            "requires the Python package `pygraphviz` (as well as the system package `graphviz`) - make sure both are installed ",
-            "http://pygraphviz.github.io/",
-        )
-
-    if format not in ["png", "svg"]:
-        raise ValueError(f"Unsupported output format: '{format}'")
-
-    G = calculate_dependency_graph(query_obj, analyse=analyse)
-    A = nx.nx_agraph.to_agraph(G)
-    s = BytesIO()
-    A.draw(s, format=format, prog="dot")
-
-    if format == "png":
-        result = Image(s.getvalue(), width=width, height=height)
-    elif format == "svg":
-        if width is not None or height is not None:  # pragma: no cover
-            logger.warning(
-                "The arguments 'width' and 'height' are not supported with format='svg'."
-            )
-        result = SVG(s.getvalue().decode("utf8"))
-    else:
-        raise ValueError(f"Unsupported output format: '{format}'")
-
-    return result
-
-
-def store_queries_in_order(dependency_graph: nx.DiGraph) -> Dict[str, "Future"]:
-    """
-    Execute queries in an order that ensures each query store is triggered after its dependencies.
-
-    Parameters
-    ----------
-    dependency_graph : networkx.DiGraph
-        Dependency graph of query objects to be stored
-    
-    Returns
-    -------
-    dict
-        Mapping from query nodes to Future objects representing the store tasks
-    """
-    ordered_list_of_queries = list(nx.topological_sort(dependency_graph))[::-1]
-    logger.debug(f"Storing queries with IDs: {ordered_list_of_queries}")
-    store_futures = {}
-    for query in ordered_list_of_queries:
-        try:
-            store_futures[query] = dependency_graph.nodes[query]["query_object"].store()
-        except UnstorableQueryError:
-            # Some queries cannot be stored
-            pass
-    return store_futures
