+++ flowmachine/tests/test_query.py
@@ -14,7 +14,6 @@ from sqlalchemy.exc import ProgrammingError
 from flowmachine.core import make_spatial_unit
 from flowmachine.core.query import Query
 from flowmachine.features import daily_location
-from flowmachine.core.dummy_query import DummyQuery
 
 
 def test_bad_sql_logged_and_raised(caplog):
@@ -157,45 +156,3 @@ def test_query_formatting():
         ValueError, match="Format string contains invalid query attribute: 'foo'"
     ):
         format(dl, "query_id,foo")
-
-
-def test_unstored_dependencies_graph():
-    """
-    Test that the _unstored_dependencies_graph method returns the correct graph in an example case.
-    """
-    # Create dummy queries with dependency structure
-    #
-    #           5:unstored
-    #            /       \
-    #       3:stored    4:unstored
-    #      /       \     /
-    # 1:unstored   2:unstored
-    #
-    # Note: we add a string parameter to each query so that they have different query IDs
-    dummy1 = DummyQuery(dummy_param=["dummy1"])
-    dummy2 = DummyQuery(dummy_param=["dummy2"])
-    dummy3 = DummyQuery(dummy_param=["dummy3", dummy1, dummy2])
-    dummy4 = DummyQuery(dummy_param=["dummy4", dummy2])
-    dummy5 = DummyQuery(dummy_param=["dummy5", dummy3, dummy4])
-    dummy3.store()
-
-    expected_query_nodes = [dummy2, dummy4]
-    graph = dummy5._unstored_dependencies_graph()
-    assert not any(dict(graph.nodes(data="stored")).values())
-    assert len(graph) == len(expected_query_nodes)
-    for query in expected_query_nodes:
-        assert f"x{query.md5}" in graph.nodes()
-        assert graph.nodes[f"x{query.md5}"]["query_object"].md5 == query.md5
-
-
-def test_unstored_dependencies_graph_for_stored_query():
-    """
-    Test that the unstored dependencies graph for a stored query is empty.
-    """
-    dummy1 = DummyQuery(dummy_param=["dummy1"])
-    dummy2 = DummyQuery(dummy_param=["dummy2"])
-    dummy3 = DummyQuery(dummy_param=["dummy3", dummy1, dummy2])
-    dummy3.store()
-
-    graph = dummy3._unstored_dependencies_graph()
-    assert len(graph) == 0
