+++ integration_tests/tests/flowmachine_server_tests/test_action_run_query.py
@@ -5,6 +5,7 @@ import os
 from flowmachine.core.cache import reset_cache
 from flowmachine.core.server.utils import send_zmq_message_and_receive_reply
 from flowmachine.core import make_spatial_unit
+from flowmachine.core.dependency_graph import unstored_dependencies_graph
 from flowmachine.features.utilities.spatial_aggregates import SpatialAggregate
 from flowmachine.features import daily_location
 from .helpers import cache_schema_is_empty, get_cache_tables, poll_until_done
@@ -128,7 +129,7 @@ async def test_cache_content(
     # Get list of tables that should be cached
     expected_cache_tables = [q.table_name]
     if "false" == os.getenv("FLOWMACHINE_SERVER_DISABLE_DEPENDENCY_CACHING"):
-        dependencies = q._unstored_dependencies_graph()
+        dependencies = unstored_dependencies_graph(q)
         for node, query_obj in dependencies.nodes(data="query_object"):
             try:
                 schema, table_name = query_obj.fully_qualified_table_name.split(".")
